--- .idea/.sourcegraph/cody_settings.json
@@ -1,45 +1,0 @@
-{
-    "cody" : {
-        "advanced" : {
-            "agent" : {
-                "capabilities" : {
-                    "storage" : true
-                },
-                "extension" : {
-                    "version" : "7.88.0"
-                },
-                "ide" : {
-                    "name" : "JetBrains",
-                    "productCode" : 12,
-                    "version" : "GO-251.23774.430"
-                },
-                "running" : true
-            },
-            "hasNativeWebview" : true
-        },
-        "autocomplete" : {
-            "advanced" : {
-                "model" : null,
-                "provider" : null
-            },
-            "enabled" : true
-        },
-        "codebase" : null,
-        "customHeaders" : {},
-        "debug" : {
-            "verbose" : false
-        },
-        "experimental" : {
-            "foldingRanges" : "indentation-based",
-            "tracing" : false
-        },
-        "serverEndpoint" : null,
-        "suggestions" : {
-            "mode" : "autocomplete"
-        },
-        "telemetry" : {
-            "clientName" : null,
-            "level" : "agent"
-        }
-    }
-}
