--- .idea/material_theme_project_new.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="MaterialThemeProjectNewConfig">
-    <option name="metadata">
-      <MTProjectMetadataState>
-        <option name="userId" value="eaad4c4:195b4bb48c6:-7fdd" />
-      </MTProjectMetadataState>
-    </option>
-  </component>
-</project>
